pyproject.toml
[tool.poetry]
name = "junior"
version = "0.1.0"
description = ""
authors = ["Alexey Lebedev <alexey.lebedev@fogstream.ru>"]

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1"
flask-restplus = "^0.13.0"
SQLAlchemy = "^1.3"
python-dotenv = "^0.10.3"
flask-sqlalchemy = "^2.4"
psycopg2-binary = "^2.8"
Flask-Migrate = "^2.5"
flask-script = "^2.0"
flask-bcrypt = "^0.7.1"
flask-admin = "^1.5"
Flask-Session = "^0.3.1"
WTForms = "^2.2"
flask-dance = "^2.2"
pyyaml = "^5.1"
flask-sessionstore = "^0.4.5"
gunicorn = "^19.9"
marshmallow = "^3.2"

[tool.poetry.dev-dependencies]
Flask-Testing = "^0.7.1"
pytest = "^5.1"
faker = "^2.0"
wemake-python-styleguide = "^0.12.5"
toml = "^0.10.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
